requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
reportlab==4.0.7
Pillow==10.1.0
python-docx==1.1.0
aiofiles==23.2.1
jinja2==3.1.2
PyMuPDF==1.23.14 